.claude/skills/workout-chart/requirements.txt
psycopg2-binary>=2.9.0
matplotlib>=3.5.0
python-dotenv>=0.19.0
pandas>=1.3.0
